Move parser type tests into a types folder
