+++ tests/test_parser.py
@@ -85,38 +85,6 @@ def sum(frm: num, to: num) -> num:
     print('Passed more complex repeater with offset test')
 
 
-def test_array_accessor():
-    array_accessor = """
-def test_array(x: num, y: num, z: num, w: num) -> num:
-    a: num[4]
-    a[0] = x
-    a[1] = y
-    a[2] = z
-    a[3] = w
-    return a[0] * 1000 + a[1] * 100 + a[2] * 10 + a[3]
-    """
-
-    c = get_contract_with_gas_estimation(array_accessor)
-    assert c.test_array(2, 7, 1, 8) == 2718
-    print('Passed basic array accessor test')
-
-
-def test_two_d_array_accessor():
-    two_d_array_accessor = """
-def test_array(x: num, y: num, z: num, w: num) -> num:
-    a: num[2][2]
-    a[0][0] = x
-    a[0][1] = y
-    a[1][0] = z
-    a[1][1] = w
-    return a[0][0] * 1000 + a[0][1] * 100 + a[1][0] * 10 + a[1][1]
-    """
-
-    c = get_contract_with_gas_estimation(two_d_array_accessor)
-    assert c.test_array(2, 7, 1, 8) == 2718
-    print('Passed complex array accessor test')
-
-
 def test_digit_reverser():
     digit_reverser = """
 
@@ -241,110 +209,6 @@ def fop():
     with pytest.raises(t.TransactionFailed):
         c.fop()
 
-def test_decimal_test():
-    decimal_test = """
-def foo() -> num:
-    return(floor(999.0))
-
-def fop() -> num:
-    return(floor(333.0 + 666.0))
-
-def foq() -> num:
-    return(floor(1332.1 - 333.1))
-
-def bar() -> num:
-    return(floor(27.0 * 37.0))
-
-def baz() -> num:
-    x = 27.0
-    return(floor(x * 37.0))
-
-def baffle() -> num:
-    return(floor(27.0 * 37))
-
-def mok() -> num:
-    return(floor(999999.0 / 7.0 / 11.0 / 13.0))
-
-def mol() -> num:
-    return(floor(499.5 / 0.5))
-
-def mom() -> num:
-    return(floor(1498.5 / 1.5))
-
-def mon() -> num:
-    return(floor(2997.0 / 3))
-
-def moo() -> num:
-    return(floor(2997 / 3.0))
-
-def foom() -> num:
-    return(floor(1999.0 % 1000.0))
-
-def foon() -> num:
-    return(floor(1999.0 % 1000))
-
-def foop() -> num:
-    return(floor(1999 % 1000.0))
-    """
-
-    c = get_contract(decimal_test)
-    pre_txs = len(s.head_state.receipts)
-    assert c.foo() == 999
-    assert c.fop() == 999
-    assert c.foq() == 999
-    assert c.bar() == 999
-    assert c.baz() == 999
-    assert c.baffle() == 999
-    assert c.mok() == 999
-    assert c.mol() == 999
-    assert c.mom() == 999
-    assert c.mon() == 999
-    assert c.moo() == 999
-    assert c.foom() == 999
-    assert c.foon() == 999
-    assert c.foop() == 999
-    post_txs = len(s.head_state.receipts)
-
-    print('Passed basic addition, subtraction and multiplication tests')
-    check_gas(decimal_test, num_txs=(post_txs - pre_txs))
-
-
-def test_harder_decimal_test():
-    harder_decimal_test = """
-def phooey(inp: decimal) -> decimal:
-    x = 10000.0
-    for i in range(4):
-        x = x * inp
-    return x
-
-def arg(inp: decimal) -> decimal:
-    return inp
-
-def garg() -> decimal:
-    x = 4.5
-    x *= 1.5
-    return x
-
-def harg() -> decimal:
-    x = 4.5
-    x *= 2
-    return x
-
-def iarg() -> wei_value:
-    x = as_wei_value(7, wei)
-    x *= 2
-    return x
-    """
-    c = get_contract(harder_decimal_test)
-    assert c.phooey(1.2) == 20736.0
-    assert c.phooey(-1.2) == 20736.0
-    assert c.arg(-3.7) == -3.7
-    assert c.arg(3.7) == 3.7
-    assert c.garg() == 6.75
-    assert c.harg() == 9.0
-    assert c.iarg() == 14
-
-    print('Passed fractional multiplication test')
 
 def test_break_test():
     break_test = """
@@ -926,153 +790,6 @@ def refund():
     print('Passed second composite crowdfund test')
 
 
-def test_test_bytes():
-    test_bytes = """
-def foo(x: bytes <= 100) -> bytes <= 100:
-    return x
-    """
-
-    c = get_contract(test_bytes)
-    moo_result = c.foo(b'cow')
-    assert moo_result == b'cow'
-
-    print('Passed basic bytes test')
-
-    assert c.foo(b'\x35' * 100) == b'\x35' * 100
-
-    print('Passed max-length bytes test')
-
-    try:
-        c.foo(b'\x35' * 101)
-        assert False
-    except:
-        pass
-
-    print('Passed input-too-long test')
-
-
-def test_test_bytes2():
-    test_bytes2 = """
-def foo(x: bytes <= 100) -> bytes <= 100:
-    y = x
-    return y
-    """
-
-    c = get_contract(test_bytes2)
-    assert c.foo(b'cow') == b'cow'
-    assert c.foo(b'') == b''
-    assert c.foo(b'\x35' * 63) == b'\x35' * 63
-    assert c.foo(b'\x35' * 64) == b'\x35' * 64
-    assert c.foo(b'\x35' * 65) == b'\x35' * 65
-
-    print('Passed string copying test')
-
-
-def test_test_bytes3():
-    test_bytes3 = """
-x: num
-maa: bytes <= 60
-y: num
-
-def __init__():
-    self.x = 27
-    self.y = 37
-
-def set_maa(inp: bytes <= 60):
-    self.maa = inp
-
-def set_maa2(inp: bytes <= 60):
-    ay = inp
-    self.maa = ay
-
-def get_maa() -> bytes <= 60:
-    return self.maa
-
-def get_maa2() -> bytes <= 60:
-    ay = self.maa
-    return ay
-
-def get_xy() -> num:
-    return self.x * self.y
-    """
-
-    c = get_contract(test_bytes3)
-    c.set_maa(b"pig")
-    assert c.get_maa() == b"pig"
-    assert c.get_maa2() == b"pig"
-    c.set_maa2(b"")
-    assert c.get_maa() == b""
-    assert c.get_maa2() == b""
-    c.set_maa(b"\x44" * 60)
-    assert c.get_maa() == b"\x44" * 60
-    assert c.get_maa2() == b"\x44" * 60
-    c.set_maa2(b"mongoose")
-    assert c.get_maa() == b"mongoose"
-    assert c.get_xy() == 999
-
-    print('Passed advanced string copying test')
-
-
-def test_test_bytes4():
-    test_bytes4 = """
-a: bytes <= 60
-def foo(inp: bytes <= 60) -> bytes <= 60:
-    self.a = inp
-    self.a = None
-    return self.a
-
-def bar(inp: bytes <= 60) -> bytes <= 60:
-    b = inp
-    b = None
-    return b
-    """
-
-    c = get_contract(test_bytes4)
-    assert c.foo() == b"", c.foo()
-    assert c.bar() == b""
-
-    print('Passed string deleting test')
-
-
-def test_test_bytes5():
-    test_bytes5 = """
-g: {a: bytes <= 50, b: bytes <= 50}
-
-def foo(inp1: bytes <= 40, inp2: bytes <= 45):
-    self.g = {a: inp1, b: inp2}
-
-def check1() -> bytes <= 50:
-    return self.g.a
-
-def check2() -> bytes <= 50:
-    return self.g.b
-
-def bar(inp1: bytes <= 40, inp2: bytes <= 45) -> bytes <= 50:
-    h = {a: inp1, b: inp2}
-    return h.a
-
-def bat(inp1: bytes <= 40, inp2: bytes <= 45) -> bytes <= 50:
-    h = {a: inp1, b: inp2}
-    return h.b
-
-def quz(inp1: bytes <= 40, inp2: bytes <= 45):
-    h = {a: inp1, b: inp2}
-    self.g = h
-    """
-
-    c = get_contract(test_bytes5)
-    c.foo(b"cow", b"horse")
-    assert c.check1() == b"cow"
-    assert c.check2() == b"horse"
-    assert c.bar(b"pig", b"moose") == b"pig"
-    assert c.bat(b"pig", b"moose") == b"moose"
-    c.quz(b"badminton", b"fluffysheep")
-    assert c.check1() == b"badminton"
-    assert c.check2() == b"fluffysheep"
-
-    print('Passed string struct test')
-
-
 def test_test_slice():
     test_slice = """
 def foo(inp1: bytes <= 10) -> bytes <= 3:
@@ -1239,77 +956,6 @@ def krazykonkat(z: bytes <= 10) -> bytes <= 25:
     assert c.krazykonkat(b"moose") == b'cow horse moose'
 
     print('Passed third concat test')
-
-
-def test_string_literal_code():
-    string_literal_code = """
-def foo() -> bytes <= 5:
-    return "horse"
-
-def bar() -> bytes <= 10:
-    return concat("b", "a", "d", "m", "i", "", "nton")
-
-def baz() -> bytes <= 40:
-    return concat("0123456789012345678901234567890", "12")
-
-def baz2() -> bytes <= 40:
-    return concat("01234567890123456789012345678901", "12")
-
-def baz3() -> bytes <= 40:
-    return concat("0123456789012345678901234567890", "1")
-
-def baz4() -> bytes <= 100:
-    return concat("01234567890123456789012345678901234567890123456789",
-                  "01234567890123456789012345678901234567890123456789")
-    """
-
-    c = get_contract(string_literal_code)
-    assert c.foo() == b"horse"
-    assert c.bar() == b"badminton"
-    assert c.baz() == b"012345678901234567890123456789012"
-    assert c.baz2() == b"0123456789012345678901234567890112"
-    assert c.baz3() == b"01234567890123456789012345678901"
-    assert c.baz4() == b"0123456789" * 10
-
-    print("Passed string literal test")
-
-
-def test_kode():
-    for i in range(95, 96, 97):
-        kode = """
-moo: bytes <= 100
-
-def foo(s: num, L: num) -> bytes <= 100:
-        x = 27
-        r = slice("%s", start=s, len=L)
-        y = 37
-        if x * y == 999:
-            return r
-
-def bar(s: num, L: num) -> bytes <= 100:
-        self.moo = "%s"
-        x = 27
-        r = slice(self.moo, start=s, len=L)
-        y = 37
-        if x * y == 999:
-            return r
-
-def baz(s: num, L: num) -> bytes <= 100:
-        x = 27
-        self.moo = slice("%s", start=s, len=L)
-        y = 37
-        if x * y == 999:
-            return self.moo
-        """ % (("c" * i), ("c" * i), ("c" * i))
-        c = get_contract(kode)
-        for e in range(63, 64, 65):
-            for _s in range(31, 32, 33):
-                o1 = c.foo(_s, e - _s)
-                o2 = c.bar(_s, e - _s)
-                o3 = c.baz(_s, e - _s)
-                assert o1 == o2 == o3 == b"c" * (e - _s), (i, _s, e - _s, o1, o2, o3) 
-
-    print("Passed string literal splicing fuzz-test")
 
 
 def test_hash_code():
@@ -1482,35 +1128,6 @@ def fivetimes(inp: bytes32) -> bytes <= 160:
     print("Passed concat bytes32 test")
 
 
-def test_test_wei():
-    test_wei = """
-def return_2_finney() -> wei_value:
-    return as_wei_value(2, finney)
-
-def return_3_finney() -> wei_value:
-    return as_wei_value(2 + 1, finney)
-
-def return_2p5_ether() -> wei_value:
-    return as_wei_value(2.5, ether)
-
-def return_3p5_ether() -> wei_value:
-    return as_wei_value(2.5 + 1, ether)
-
-def return_2pow64_wei() -> wei_value:
-    return as_wei_value(18446744.073709551616, szabo)
-    """
-
-    c = get_contract_with_gas_estimation(test_wei)
-
-    assert c.return_2_finney() == 2 * 10**15
-    assert c.return_3_finney() == 3 * 10**15, c.return_3_finney()
-    assert c.return_2p5_ether() == 2.5 * 10**18
-    assert c.return_3p5_ether() == 3.5 * 10**18
-    assert c.return_2pow64_wei() == 2**64
-
-    print("Passed wei value literals test")
-
-
 def test_caller_code():
     caller_code = """
 def foo() -> bytes <= 7:
@@ -1571,43 +1188,6 @@ def foq(inp: bytes <= 32) -> address:
     assert not success
 
     print('Passed extract32 test')
-
-
-def test_bytes_to_num_code():
-    bytes_to_num_code = """
-def foo(x: bytes <= 32) -> num:
-    return bytes_to_num(x)
-    """
-
-    c = get_contract(bytes_to_num_code)
-    assert c.foo(b"") == 0
-    try:
-        c.foo(b"\x00")
-        success = True
-    except:
-        success = False
-    assert not success
-    assert c.foo(b"\x01") == 1
-    try:
-        c.foo(b"\x00\x01")
-        success = True
-    except:
-        success = False
-    assert not success
-    assert c.foo(b"\x01\x00") == 256
-    assert c.foo(b"\x01\x00\x00\x00\x01") == 4294967297
-    assert c.foo(b"\xff" * 32) == -1
-    try:
-        c.foo(b"\x80" + b"\xff" * 31)
-        success = True
-    except:
-        success = False
-    try:
-        c.foo(b"\x01" * 33)
-        success = True
-    except:
-        success = False
-    print('Passed bytes_to_num tests')
 
 
 def test_rlp_decoder_code():
@@ -1909,57 +1489,6 @@ def _shift(x: num256, y: num) -> num256:
     print("Passed bitwise operation tests")
 
 
-def test_num256_code():
-    num256_code = """
-def _num256_add(x: num256, y: num256) -> num256:
-    return num256_add(x, y)
-
-def _num256_sub(x: num256, y: num256) -> num256:
-    return num256_sub(x, y)
-
-def _num256_mul(x: num256, y: num256) -> num256:
-    return num256_mul(x, y)
-
-def _num256_div(x: num256, y: num256) -> num256:
-    return num256_div(x, y)
-
-def _num256_gt(x: num256, y: num256) -> bool:
-    return num256_gt(x, y)
-
-def _num256_ge(x: num256, y: num256) -> bool:
-    return num256_ge(x, y)
-
-def _num256_lt(x: num256, y: num256) -> bool:
-    return num256_lt(x, y)
-
-def _num256_le(x: num256, y: num256) -> bool:
-    return num256_le(x, y)
-    """
-
-    c = get_contract_with_gas_estimation(num256_code)
-    x = 126416208461208640982146408124
-    y = 7128468721412412459
-
-    assert c._num256_add(x, y) == x + y
-    assert c._num256_sub(x, y) == x - y
-    assert c._num256_sub(y, x) == 2**256 + y - x
-    assert c._num256_mul(x, y) == x * y
-    assert c._num256_mul(2**128, 2**128) == 0
-    assert c._num256_div(x, y) == x // y
-    assert c._num256_div(y, x) == 0
-    assert c._num256_gt(x, y) is True
-    assert c._num256_ge(x, y) is True
-    assert c._num256_le(x, y) is False
-    assert c._num256_lt(x, y) is False
-    assert c._num256_gt(x, x) is False
-    assert c._num256_ge(x, x) is True
-    assert c._num256_le(x, x) is True
-    assert c._num256_lt(x, x) is False
-    assert c._num256_lt(y, x) is True
-
-    print("Passed num256 operation tests")
-
-
 def test_selfcall_code():
     selfcall_code = """
 def foo() -> num:
@@ -2173,107 +1702,6 @@ def create_and_return_forwarder(inp: address) -> address:
     assert not success
 
     print('Passed forwarder exception test')
-
-
-def test_list_tester_code():
-    list_tester_code = """
-z: num[3]
-z2: num[2][2]
-z3: num[2]
-
-def foo(x: num[3]) -> num:
-    return x[0] + x[1] + x[2]
-
-def goo(x: num[2][2]) -> num:
-    return x[0][0] + x[0][1] + x[1][0] * 10 + x[1][1] * 10
-
-def hoo(x: num[3]) -> num:
-    y = x
-    return y[0] + x[1] + y[2]
-
-def joo(x: num[2][2]) -> num:
-    y = x
-    y2 = x[1]
-    return y[0][0] + y[0][1] + y2[0] * 10 + y2[1] * 10
-
-def koo(x: num[3]) -> num:
-    self.z = x
-    return self.z[0] + x[1] + self.z[2]
-
-def loo(x: num[2][2]) -> num:
-    self.z2 = x
-    self.z3 = x[1]
-    return self.z2[0][0] + self.z2[0][1] + self.z3[0] * 10 + self.z3[1] * 10
-    """
-
-    c = get_contract(list_tester_code)
-    assert c.foo([3,4,5]) == 12
-    assert c.goo([[1,2],[3,4]]) == 73
-    assert c.hoo([3,4,5]) == 12
-    assert c.joo([[1,2],[3,4]]) == 73
-    assert c.koo([3,4,5]) == 12
-    assert c.loo([[1,2],[3,4]]) == 73
-    print("Passed list tests")
-
-
-def test_list_output_tester_code():
-    list_output_tester_code = """
-z: num[2]
-
-def foo() -> num[2]:
-    return [3, 5]
-
-def goo() -> num[2]:
-    x = [3, 5]
-    return x
-
-def hoo() -> num[2]:
-    self.z = [3, 5]
-    return self.z
-
-def joo() -> num[2]:
-    self.z = [3, 5]
-    x = self.z
-    return x
-
-def koo() -> num[2][2]:
-    return [[1,2],[3,4]]
-
-def loo() -> num[2][2]:
-    x = [[1,2],[3,4]]
-    return x
-
-def moo() -> num[2][2]:
-    x = [1,2]
-    return [x,[3,4]]
-
-def noo(inp: num[2]) -> num[2]:
-    return inp
-
-def poo(inp: num[2][2]) -> num[2][2]:
-    return inp
-
-def qoo(inp: num[2]) -> num[2][2]:
-    return [inp,[3,4]]
-
-def roo(inp: num[2]) -> decimal[2][2]:
-    return [inp,[3,4]]
-    """
-
-    c = get_contract(list_output_tester_code)
-    assert c.foo() == [3,5]
-    assert c.goo() == [3,5]
-    assert c.hoo() == [3,5]
-    assert c.joo() == [3,5]
-    assert c.koo() == [[1,2],[3,4]]
-    assert c.loo() == [[1,2],[3,4]]
-    assert c.moo() == [[1,2],[3,4]]
-    assert c.noo([3,5]) == [3,5]
-    assert c.poo([[1,2],[3,4]]) == [[1,2],[3,4]]
-    assert c.qoo([1,2]) == [[1,2],[3,4]]
-    assert c.roo([1,2]) == [[1.0,2.0],[3.0,4.0]]
-
-    print("Passed list output tests")
 
 
 def test_internal_test():
